Moved logging outside of ml_worker module
