+++ giskard/models/base/model.py
@@ -24,10 +24,10 @@ from ...core.core import ModelMeta, ModelType, SupportedModelTypes
 from ...core.validation import configured_validate_arguments
 from ...datasets.base import Dataset
 from ...exceptions.giskard_exception import GiskardException, python_env_exception_helper
-from ...ml_worker.utils.logging import Timer
 from ...models.cache import ModelCache
 from ...path_utils import get_size
 from ...settings import settings
+from ...utils.logging import Timer
 
 META_FILENAME = "giskard-model-meta.yaml"
