+++ giskard/models/model_explanation.py
@@ -1,15 +1,14 @@
-from typing import Any, Callable, Dict, List
-
 import logging
 import warnings
 
 import numpy as np
 import pandas as pd
+from typing import Any, Callable, Dict, List
 
 from giskard.core.errors import GiskardImportError
 from giskard.datasets.base import Dataset
-from giskard.ml_worker.utils.logging import timer
 from giskard.models.base import BaseModel
+from giskard.utils.logging import timer
 
 warnings.filterwarnings("ignore", message=".*The 'nopython' keyword.*")
 logger = logging.getLogger(__name__)
