+++ tests/fixtures/diabetes__regression.py
@@ -5,8 +5,8 @@ from sklearn.metrics import mean_squared_error
 
 from giskard.core.core import SupportedModelTypes
 from giskard.datasets.base import Dataset
-from giskard.ml_worker.utils.logging import Timer
 from giskard.models.sklearn import SKLearnModel
+from giskard.utils.logging import Timer
 
 
 @pytest.fixture(scope="session")
