+++ tests/fixtures/german_credit_scoring.py
@@ -12,9 +12,9 @@ from sklearn.preprocessing import OneHotEncoder, StandardScaler
 
 from giskard.core.core import SupportedModelTypes
 from giskard.datasets.base import Dataset
-from giskard.ml_worker.utils.logging import Timer
 from giskard.models.catboost import CatboostModel
 from giskard.models.sklearn import SKLearnModel
+from giskard.utils.logging import Timer
 from tests import path
 
 logger = logging.getLogger(__name__)
